requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
